# Fix build issues by removing FCollada's bundled libxml2 On Windows, dynamically link to our standard libraries/libxml2/... implementation.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7035 3db68df2-c116-0410-a063-a993310a9797

+++ libraries/fcollada/src/FCollada/FUtils/FUXmlDocument.cpp
@@ -36,7 +36,7 @@ FUXmlDocument::FUXmlDocument(FUFileManager* manager, const fchar* _filename, boo
 			file->Close();
 
 			// Open the given XML file.
-			xmlDocument = xmlParseMemory((const char*) fileData, fileLength);
+			xmlDocument = xmlParseMemory((const char*) fileData, (int)fileLength);
 			SAFE_DELETE_ARRAY(fileData);
 		}
 		SAFE_DELETE(file);
@@ -64,7 +64,7 @@ FUXmlDocument::FUXmlDocument(const char* data, size_t length)
 	}
 
 	// Open the given XML file.
-	xmlDocument = xmlParseMemory(data, length);
+	xmlDocument = xmlParseMemory(data, (int)length);
 }
 
 FUXmlDocument::~FUXmlDocument()
